Moves Backbone class definition to a new backbone_base module in order to avoid the circular import error.
This happens when model.resnet and model.vgg modules try to import the Backbone class from model.Backbone, since model.Backbone also tries to import from model.resnet and model.vgg, it creates a circular import problem.

Since the Backbone class is not used in other parts of the model.Backbone module, it can be moved to a new  module (model.backbone_base) that does not import anything from model.resnet and model.vgg. Thus, by making model.resnet and model.vgg import from model.backbone_base, we solve the circular import problem.
